Remove the use of six from PBR
Outside of testing the only real place six was used was to import
configparser in a backward and forward compatible manner. We can do
conditional imports ourselves and drop the user of six from pbr which
drops a runtime requirement.

In testing we used six for configparser and unicode strings. We fix
configparser as in pbr proper and simply use u'' strings where six.u()
was previously used.

Change-Id: I3b38c9a9c3c1a8baf38993b459ea72715b91776a