Use sysconfig for sitedir path in test_wsgi in py3
Use `sysconfig.get_path()` to obtain the correct site-packages directory
path in `test_wsgi` for Python 3, instead of attempting to guess it
based on Python version.  This fixes the test on PyPy3.10, and seems to
be correct for CPython 3.  While locally the call seems to return the
correct result in Python 2.7 as well, the test suite seems to fail with
it, so I've just kept the old code for older Python versions.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Partial-Bug: 2097427
Change-Id: I5c152a98fd371dfb195643f0f5640cf1ffe0ba31