tests: Skip tests that break on newer setuptools
These are all broken on setuptools >= v80.0.0 due to the removal of
easy_install there. This solution is at most a stop-gap and we have a
bigger issue to resolve here - namely figuring out how to support these
once distros start packaging newer setuptools. However, it at least
unblocks the gate for us.

We also need to fix another issue, ensuring that we normalize
constraints as we do package names. This fixes issue with glance-store,
which is listed as 'glance_store' since [1].

[1] https://review.opendev.org/c/openstack/requirements/+/942508/

Change-Id: I599cb05706753c89f0ba74ee17446e62abd79b5d
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>