Add black
Ahead of major surgery. You might ask why Black rather than Ruff? We
still care about Python 2.7 for now, meaning we need to explicitly mark
unicode strings as such with the 'u' prefix (and 'b' for byte strings).
Ruff has the 'ruff.format.quote-style = preserve' option that is
supposed to retain this, but it doesn't preserve the prefixes. Black
does and is therefore used for now.

Change-Id: Ia5f447f1dd27486ac540ce10723ab7002d8ec39e
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>