Tweak loading distribution version
These appear to be functionally identical under a typical install, and
if there is a difference then I can't figure it out. Making this changes
makes the application of a compat layer in a future commit somewhat
easier. We keep it to a separate commit in case this comes back to bite
us, however.

Change-Id: Ib6715fd74025e2ec194ef6258c1433b697498b93
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>