Remove use of easy_install
Vendor the functionality we were relying on from setuptools v67.8.0 [1]
(the last version to provide all the deprecated features). Remove
everything else.

This also fixes a bug on the Windows code path, which was using an API
that was removed in setuptools v68.0.0 [2]. I suspect we had not
triggered this since no one is installing services (and therefore taking
advantage of the 'wsgi_scripts' functionality) there.

[1] https://github.com/pypa/setuptools/blob/v67.8.0/setuptools/command/easy_install.py
[2] https://github.com/pypa/setuptools/commit/e96f6a3ee3b0df47b4731a200695bd6151f64111

Change-Id: Ic9b022c5166ec1df4122e7c5962ded9ca8e934d9
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>