Pin zuul ansible, pyperclip versions
zuul recently moved to Ansible 11, which does not support Python < 3.8.
We need to pin the Ansible version on the older jobs.

Also pin pyperclip, recent versions of which claim to support Python 2.7
but in practice do not [1].

Finally, we readd (and reword) a note about dependency order that was
incorrectly removed previously since it still applies to the versions of
pip found in older Python 2.7 environments (like Ubuntu 18.04).

[1] https://github.com/asweigart/pyperclip/issues/285

Change-Id: I74b43ad580c6514f06d5af64cb969c02256b5caa
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>